requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
google-generativeai==0.8.5
flask-cors==4.0.0
fpdf==1.7.2
